+++ Emulator/Cartridges/CustomCartridges/ActionReplay.h
@@ -46,7 +46,7 @@ public:
     // Handling buttons
     //
 
-    unsigned numButtons() override { return 2; }
+    long numButtons() override { return 2; }
     const char *getButtonTitle(unsigned nr) override;
     void pressButton(unsigned nr) override;
     void releaseButton(unsigned nr) override;
@@ -99,7 +99,7 @@ public:
     // Handling buttons
     //
     
-    unsigned numButtons() override { return 2; }
+    long numButtons() override { return 2; }
     const char *getButtonTitle(unsigned nr) override;
     void pressButton(unsigned nr) override;
     void releaseButton(unsigned nr) override;
