+++ Emulator/Cartridges/CustomCartridges/Expert.h
@@ -93,7 +93,7 @@ private:
     // Operating buttons
     //
     
-    unsigned numButtons() override { return 2; }
+    long numButtons() override { return 2; }
     const char *getButtonTitle(unsigned nr) override;
     void pressButton(unsigned nr) override;
